+++ exercicios/para-casa/gabrielalima/projeto.py
@@ -1,0 +1,320 @@
+import print_reprograma as pr
+from datetime import datetime
+
+produtos = [
+    {'codigo': 1, 'nome': 'Introdução à Computação para Alunos Inteligentes - Volume 1', 'valor': 10.00},
+    {'codigo': 2, 'nome': 'Introdução à Computação para Alunos Inteligentes - Volume 2', 'valor': 10.00},
+    {'codigo': 3, 'nome': 'Linguagens de Programação Web para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 4, 'nome': 'Tecnologias de Scripting para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 5, 'nome': 'Desenvolvimento Web com SQL para Alunos Inteligentes', 'valor': 15.00},
+    {'codigo': 6, 'nome': 'Desenvolvimento Web com C++ para Alunos Inteligentes', 'valor': 15.00},
+    {'codigo': 7, 'nome': 'Desenvolvimento Web com Java para Alunos Inteligentes', 'valor': 15.00},
+    {'codigo': 8, 'nome': 'Desenvolvimento Web com .Net para Alunos Inteligentes', 'valor': 15.00},
+    {'codigo': 9, 'nome': 'Hardware e Redes de Computadores para Alunos Inteligentes - Volume 1', 'valor': 15.00},
+    {'codigo': 10, 'nome': 'Hardware e Redes de Computadores para Alunos Inteligentes - Volume 2', 'valor': 15.00},
+    {'codigo': 11, 'nome': 'Hardware e Redes de Computadores para Alunos Inteligentes - Volume 3', 'valor': 15.00},
+    {'codigo': 12, 'nome': 'Hardware e Redes de Computadores para Alunos Inteligentes - Volume 4', 'valor': 15.00},
+    {'codigo': 13, 'nome': 'Dominando o Inglês para Alunos Inteligentes', 'valor': 5.00},
+    {'codigo': 14, 'nome': 'Comunicação Empresarial para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 15, 'nome': 'Desenvolvimento Pessoal para Alunos Inteligentes - Livro 1', 'valor': 15.00},
+    {'codigo': 16, 'nome': 'Desenvolvimento Pessoal para Alunos Inteligentes - Livro 2', 'valor': 15.00},
+    {'codigo': 17, 'nome': 'Introdução à Computação para Alunos Inteligentes - Volume 1', 'valor': 10.00},
+    {'codigo': 18, 'nome': 'Introdução à Computação para Alunos Inteligentes - Volume 2', 'valor': 10.00},
+    {'codigo': 19, 'nome': 'Fundamentos da Programação para Alunos Inteligentes', 'valor': 12.00},
+    {'codigo': 20, 'nome': 'Algoritmos e Estruturas de Dados para Alunos Inteligentes', 'valor': 15.00},
+    {'codigo': 21, 'nome': 'Introdução à Engenharia de Software para Alunos Inteligentes', 'valor': 18.00},
+    {'codigo': 22, 'nome': 'Banco de Dados para Alunos Inteligentes', 'valor': 12.00},
+    {'codigo': 23, 'nome': 'Programação Web com Python para Alunos Inteligentes', 'valor': 15.00},
+    {'codigo': 24, 'nome': 'Programação Web com JavaScript para Alunos Inteligentes', 'valor': 15.00},
+    {'codigo': 25, 'nome': 'Introdução à Ciência da Computação para Alunos Inteligentes', 'valor': 18.00},
+    {'codigo': 26, 'nome': 'Inteligência Artificial para Alunos Inteligentes', 'valor': 20.00},
+    {'codigo': 27, 'nome': 'Redes de Computadores para Alunos Inteligentes', 'valor': 15.00},
+    {'codigo': 28, 'nome': 'Segurança da Informação para Alunos Inteligentes', 'valor': 18.00},
+    {'codigo': 29, 'nome': 'Administração de Sistemas para Alunos Inteligentes', 'valor': 15.00},
+    {'codigo': 30, 'nome': 'Gestão de Projetos de TI para Alunos Inteligentes', 'valor': 12.00},
+    {'codigo': 31, 'nome': 'Introdução à Estatística para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 32, 'nome': 'Probabilidade para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 33, 'nome': 'Matemática Discreta para Alunos Inteligentes', 'valor': 12.00},
+    {'codigo': 34, 'nome': 'Cálculo para Alunos Inteligentes', 'valor': 15.00},
+    {'codigo': 35, 'nome': 'Álgebra Linear para Alunos Inteligentes', 'valor': 12.00},
+    {'codigo': 36, 'nome': 'Geometria Analítica para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 37, 'nome': 'Física para Alunos Inteligentes', 'valor': 15.00},
+    {'codigo': 38, 'nome': 'Química para Alunos Inteligentes', 'valor': 12.00},
+    {'codigo': 39, 'nome': 'Biologia para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 40, 'nome': 'História para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 41, 'nome': 'Geografia para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 42, 'nome': 'Sociologia para Alunos Inteligentes', 'valor': 12.00},
+    {'codigo': 43, 'nome': 'Filosofia para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 44, 'nome': 'Literatura Brasileira para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 45, 'nome': 'Literatura Portuguesa para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 46, 'nome': 'Literatura Inglesa para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 47, 'nome': 'Literatura Espanhola para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 48, 'nome': 'Literatura Francesa para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 49, 'nome': 'Literatura Alemã para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 50, 'nome': 'Literatura Italiana para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 51, 'nome': 'Literatura Russa para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 52, 'nome': 'Literatura Japonesa para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 53, 'nome': 'Literatura Chinesa para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 54, 'nome': 'Arte para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 55, 'nome': 'Música para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 56, 'nome': 'Cinema para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 57, 'nome': 'Teatro para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 58, 'nome': 'Dança para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 59, 'nome': 'Fotografia para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 60, 'nome': 'Direito para Alunos Inteligentes', 'valor': 15.00},
+    {'codigo': 61, 'nome': 'Economia para Alunos Inteligentes', 'valor': 12.00},
+    {'codigo': 62, 'nome': 'Administração para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 63, 'nome': 'Contabilidade para Alunos Inteligentes', 'valor': 12.00},
+    {'codigo': 64, 'nome': 'Marketing para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 65, 'nome': 'Publicidade e Propaganda para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 66, 'nome': 'Recursos Humanos para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 67, 'nome': 'Relações Públicas para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 68, 'nome': 'Comunicação para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 69, 'nome': 'Jornalismo para Alunos Inteligentes', 'valor': 12.00},
+    {'codigo': 70, 'nome': 'Psicologia para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 71, 'nome': 'Sociologia para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 72, 'nome': 'Antropologia para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 73, 'nome': 'Ciência Política para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 74, 'nome': 'História da Arte para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 75, 'nome': 'História da Música para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 76, 'nome': 'História do Cinema para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 77, 'nome': 'História do Teatro para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 78, 'nome': 'História da Dança para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 79, 'nome': 'História da Fotografia para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 80, 'nome': 'Filosofia da Arte para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 81, 'nome': 'Filosofia da Música para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 82, 'nome': 'Filosofia do Cinema para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 83, 'nome': 'Filosofia do Teatro para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 84, 'nome': 'Filosofia da Dança para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 85, 'nome': 'Filosofia da Fotografia para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 86, 'nome': 'Ética para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 87, 'nome': 'Ética Profissional para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 88, 'nome': 'Ética e Responsabilidade Social para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 89, 'nome': 'Lógica para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 90, 'nome': 'Pensamento Crítico para Alunos Inteligentes', 'valor': 10.00},
+    {'codigo': 91, 'nome': 'Introdução ao Direito para Alunos Inteligentes', 'valor': 15.00},
+    {'codigo': 92, 'nome': 'Direito Constitucional para Alunos Inteligentes', 'valor': 15.00},
+    {'codigo': 93, 'nome': 'Direito Civil para Alunos Inteligentes', 'valor': 15.00},
+    {'codigo': 94, 'nome': 'Direito Penal para Alunos Inteligentes', 'valor': 15.00},
+    {'codigo': 95, 'nome': 'Direito Processual para Alunos Inteligentes', 'valor': 15.00},
+    {'codigo': 96, 'nome': 'Direito do Trabalho para Alunos Inteligentes', 'valor': 15.00},
+    {'codigo': 97, 'nome': 'Direito Tributário para Alunos Inteligentes', 'valor': 15.00},
+    {'codigo': 98, 'nome': 'Direito Empresarial para Alunos Inteligentes', 'valor': 15.00},
+    {'codigo': 99, 'nome': 'Direito Internacional para Alunos Inteligentes', 'valor': 15.00},
+    {'codigo': 100, 'nome': 'Direito Ambiental para Alunos Inteligentes', 'valor': 15.00},
+]
+
+def produto_codigo(codigo):
+    for produto in produtos:
+        if produto["codigo"] == codigo:
+            return produto
+
+def criar_novo_produto(produto, quantidade):
+    return {
+        "codigo": produto["codigo"],
+        "nome": produto["nome"],
+        "valor": produto["valor"],
+        "quantidade": quantidade
+    }
+
+def imprime_produto(produto):
+    pr.imprimir(str(produto['codigo']), tamanho=6, alinhar='fim', caracter='0', end='|')
+    pr.imprimir(produto['nome'], tamanho=83, caracter='.', end='|')
+    pr.imprimir(str(produto['quantidade']), tamanho=3, caracter='0', alinhar='fim', end='|')
+    pr.imprimir('R$', str(round(produto['valor'], 2)), tamanho=12, alinhar='fim', end='|')
+    pr.imprimir('R$', str(round(produto.get('valor_final', produto['valor']) * produto['quantidade'], 2)), tamanho=12, alinhar='fim')
+
+def imprime_compra(compra):
+    if len(compra) > 0: 
+        total = sum(produto['valor'] * produto['quantidade'] for produto in compra)
+        pr.imprimir('codigo', tamanho=6, alinhar='centro', end='|') 
+        pr.imprimir('produto', tamanho=83, alinhar='centro', end='|') 
+        pr.imprimir('qtd', tamanho=3, alinhar='centro', end='|') 
+        pr.imprimir('valor un.', tamanho=12, alinhar='centro', end='|') 
+        pr.imprimir('valor', tamanho=12, alinhar='centro')   
+        for produto in compra: 
+            imprime_produto(produto) 
+        pr.separador(120, caracter='-') 
+        pr.imprimir('Subtotal', tamanho=107, alinhar='fim', end='|') 
+        pr.imprimir('R$', str(round(total, 2)), tamanho=12, alinhar='fim') 
+    else: 
+        pr.imprimir('Sem itens na lista ainda', tamanho=120, alinhar='centro') 
+    pr.pular_linha() 
+    pr.pular_linha()
+
+def imprime_compra_fechada(compra, total, desconto=0):
+    subtotal = sum(produto["valor"] * produto["quantidade"] for produto in compra)
+
+    # Impressão dos detalhes da compra
+    pr.imprimir('codigo', tamanho=6, alinhar='centro', end='|')
+    pr.imprimir('produto', tamanho=83, alinhar='centro', end='|')
+    pr.imprimir('qtd', tamanho=3, alinhar='centro', end='|')
+    pr.imprimir('valor un.', tamanho=12, alinhar='centro', end='|')
+    pr.imprimir('valor', tamanho=12, alinhar='centro')
+
+    for produto in compra:
+        imprime_produto(produto)
+
+    pr.separador(120, caracter='-')
+
+    # Impressão do subtotal
+    pr.imprimir('Total', tamanho=107, alinhar='fim', end='|')
+    pr.imprimir('R$', str(round(subtotal, 2)), tamanho=12, alinhar='fim')
+
+    # Impressão do desconto, se houver
+    if desconto > 0:
+        pr.imprimir('Desconto', tamanho=107, alinhar='fim', end='|')
+        pr.imprimir('- R$', str(round(desconto, 2)), tamanho=12, alinhar='fim', cor_texto='vermelho')
+        total_final = subtotal - desconto
+    else:
+        total_final = subtotal
+
+    # Impressão do total a pagar
+    pr.imprimir('Total a pagar', tamanho=107, alinhar='fim', end='|')
+    pr.imprimir('R$', str(round(total_final, 2)), tamanho=12, alinhar='fim', cor_texto='verde negrito')
+
+    # Limpeza e formatação
+    pr.limpar_formatacao()
+    pr.pular_linha()
+    pr.pular_linha()
+
+
+def calcula_desconto_total(compras):
+    total_desconto = 0
+    for compra in compras:
+        for produto in compra['itens']:
+            if 'desconto' in produto:
+                valor_desconto = produto['valor'] * (produto['desconto'] / 100) * produto['quantidade']
+                total_desconto += valor_desconto
+    return total_desconto
+    
+def imprime_fechamento_caixa(compras):
+    pr.imprimir('Data', tamanho=89, alinhar='centro', end='|')
+    pr.imprimir('Qt.', tamanho=9, alinhar='centro', end='|')
+    pr.imprimir('valor', tamanho=20, alinhar='centro', end='|')
+    pr.imprimir('Desconto', tamanho=20, alinhar='centro')
+    total = 0
+    total_desconto = calcula_desconto_total(compras)
+    for compra in compras:
+        total_compra = sum(produto['valor'] * produto['quantidade'] for produto in compra['itens'])
+        total += total_compra
+        pr.imprimir(compra['data'].strftime("%d/%m/%Y %H:%M:%S "), tamanho=89, end='|', alinhar='fim')
+        pr.imprimir(str(len(compra['itens'])), tamanho=9, end='|', alinhar='centro')
+        pr.imprimir('R$', str(round(total_compra, 2)), tamanho=20, end='|', alinhar='fim')
+        if 'desconto' in compra and compra['desconto'] > 0:
+            pr.imprimir('R$', str(round(total_compra * (compra['desconto'] / 100), 2)), tamanho=20, alinhar='fim', cor_texto='vermelho')
+        else:
+            pr.imprimir('R$ 0', tamanho=20, alinhar='fim')  # Se não houver desconto, exibe R$ 0
+    pr.separador(120, caracter='-')
+    pr.imprimir('Total de compras do caixa', tamanho=99, alinhar='fim', end='|')
+    pr.imprimir('R$', str(round(total, 2)), tamanho=20, alinhar='fim')
+    
+
+def imprimir_cabecalho(erro):
+    pr.limpar()
+    pr.retangulo("{reprograma}\n Projeto Guiado 1\n Terminal de Vendas", sv=1, tamanho=100, margem=6, cor_texto="magenta", cor_barra="azul")
+    pr.separador(108, cor_texto="ciano")
+
+    if erro != "":
+        pr.imprimir(erro, tamanho=120, alinhar="centro", cor_texto="vermelho negrito")
+        pr.separador(100, cor_texto="ciano")
+    erro = ""
+
+def imprimir_ajuda():
+    pr.pular_linha(quantidade=2)
+    pr.imprimir('[H]   >> Ajuda com o Sistema', alinhar='centro', tamanho=120)
+    pr.imprimir('[Q]   >> Sair da Tela ou Sistema', alinhar='centro', tamanho=120)
+    pr.imprimir('[N]   >> Cria uma Nova Compra', alinhar='centro', tamanho=120)
+    pr.imprimir('[F]   >> Fechar a Compra', alinhar='centro', tamanho=120)
+    pr.imprimir('[P]   >> Confirmar que a compra foi paga', alinhar='centro', tamanho=120)
+    pr.imprimir('[nnn] >> Adicionar o código do produto à compra', alinhar='centro', tamanho=120)
+    pr.imprimir('[Xnn] >> Muda a quantidade de itens adicionado', alinhar='centro', tamanho=120)
+    pr.imprimir('[E]   >> Encerrar caixa', alinhar='centro', tamanho=120)
+    pr.pular_linha(quantidade=2)
+
+def imprimir_rodape():
+    pr.imprimir("[H] Ajuda ", "[Q] Sair ", caracter="=", tamanho=108, alinhar="fim", end=" 😈 :")
+    return input().lower()
+
+def menu():
+    opcao = ''
+    erro = ''
+    tela = ''
+    compra = []
+    compras = []
+    total = 0
+    pago = False
+    while opcao != 'q':
+        imprimir_cabecalho(erro)
+        if tela == '':
+            pr.pular_linha(quantidade=4)
+        elif tela == 'ajuda':
+            imprimir_ajuda()
+            tela = ''
+        elif tela == 'compra':
+            imprime_compra(compra)
+        elif tela == "fechar":
+            if pago:
+                total = sum(produto['valor'] * produto['quantidade'] for produto in compra)
+                desconto = input("Insira a porcentagem de desconto (0-100) para a compra, ou pressione Enter para pular: ")
+                if desconto:
+                    desconto = float(desconto)
+                    if 0 <= desconto <= 100:
+                        desconto_total = total * (desconto / 100)
+                    else:
+                        erro = 'Desconto inválido! Deve estar entre 0 e 100.'
+                        continue
+                else:
+                    desconto_total = 0
+                imprime_compra_fechada(compra, total, desconto_total)
+            else:
+                erro = 'Por favor, confirme que a compra foi paga antes de fechar.'
+        elif tela == "encerrar":
+            imprime_fechamento_caixa(compras)
+            compras = []
+            tela = ""
+            pr.pular_linha(quantidade=2)
+
+        opcao = imprimir_rodape()
+
+        if opcao == 'h':
+            tela = 'ajuda'
+        elif opcao == 'n':
+            tela = 'compra'
+        elif opcao == "f":
+            if compra and pago:
+                tela = "fechar"
+            elif not compra:
+                erro = 'Não há itens na compra para fechar.'
+            elif not pago:
+                erro = 'Por favor, confirme que a compra foi paga antes de fechar.'
+        elif opcao == "e":
+            tela = "encerrar"
+        elif "p" in opcao:
+            if compra:
+                total = sum(produto['valor'] * produto['quantidade'] for produto in compra)
+                print("\nConfirma pagamento da compra? (s/n)")
+                confirmacao = input().lower()
+                if confirmacao == 's':
+                    pago = True
+                    compras.append({"itens": compra, "total": total, "data": datetime.now()})
+                    compra = []
+                    total = 0
+                    tela = ""
+                else:
+                    erro = 'Pagamento não confirmado.'
+            else:
+                erro = 'Não há itens na compra para confirmar o pagamento.'
+        else:
+            try:
+                codigo = int(opcao)
+                produto = produto_codigo(codigo)
+                quantidade = 1
+                novo_produto_adicionado = criar_novo_produto(produto, quantidade)
+                compra.append(novo_produto_adicionado)
+            except ValueError:
+                erro = 'A opção selecionada não existe no sistema'
+
+    imprimir_rodape()
+
+menu()
